+++ lindera-core/src/token.rs
@@ -1,7 +1,9 @@
+use std::borrow::Cow;
+
 use serde::Serialize;
 
 #[derive(Serialize, Clone)]
 pub struct Token<'a> {
-    pub text: &'a str,
+    pub text: Cow<'a, str>,
     pub details: Option<Vec<String>>,
 }