+++ lindera/src/token_filter.rs
@@ -3,4 +3,6 @@ pub mod japanese_keep_tags;
 pub mod japanese_stop_tags;
 pub mod keep_words;
 pub mod length;
+pub mod lowercase;
 pub mod stop_words;
+pub mod uppercase;